Add an example of a function that returns a struct
Add an example of a function that returns a struct that is created by the function. In the example, also check the type of the struct that is returned and its values.  Note that the input type for the `type_of` function is changed to a pointer so that the instance is borrowed. Otherwise when trying to check the values of the instance after calling `type_of`, the compiler complains that the instance has been moved.